pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memoria"
readme = "README.md"
# license = {file = "LICENSE"}
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Typing :: Typed",
]
dependencies = [
    "aiofiles~=23.2.1",
    "aiohttp~=3.9.5",
    "aiosqlite~=0.20.0",
    "beautifulsoup4~=4.12.3",
    "elasticsearch[async]~=8.13.1",
    "fasthx~=0.2403.1",
    "fastapi~=0.111.0",
    "humanize~=4.9.0",
    "ijson~=3.2.3",
    "pydantic-settings~=2.2.1",
    "python-magic~=0.4.27",
    "SQLAlchemy[asyncio]~=2.0.30"
]
dynamic = ["version", "description", "authors"]

[project.urls]
Repository = "https://github.com/Sidneys1/Memoria"
"Bug Tracker" = "https://github.com/Sidneys1/Memoria/issues"

[project.entry-points.memoria]
AiohttpDownloader = "memoria.plugins.builtin.aiohttp_downloader:AiohttpDownloader"
HtmlContentFinder = "memoria.plugins.builtin.html_content_finder:HtmlContentFinder"
HtmlExtractor = "memoria.plugins.builtin.html_extractor:HtmlExtractor"

[project.optional-dependencies]
dev = [
    "hatch",
    "fastapi-cli",
    "isort",
    "mypy",
]
uvicorn = [
    "uvicorn",
]

[tool.yapf]
"COLUMN_LIMIT" = 120

[tool.hatch]

[tool.hatch.version]
path = "src/memoria/__about__.py"

[tool.hatch.metadata.hooks.custom]

[tool.hatch.build.targets.sdist]
exclude = [
     "allowlist.txt",
     "denylist.txt",
     ".vscode/",
     ".editorconfig"
]

[tool.isort]
line_length = 120
ensure_newline_before_comments = true
